Move app dispatchRouterAndPrepareRoute method to trait
